+++ src/bypp.schema.ts
@@ -8,6 +8,6 @@
  */
 
 export {
-  BeyondPaperV18Schema as BeyondPaperSchema,
-  type BeyondPaperV18 as BeyondPaper,
-} from "./schemas/bypp.v18.schema";
+  BeyondPaperV19Schema as BeyondPaperSchema,
+  type BeyondPaperV19 as BeyondPaper,
+} from "./schemas/bypp.v19.schema";